+++ CHANGELOG.md
@@ -8,9 +8,11 @@
 * Added a /Name entry to the default optional content configuration dictionary
   (needed by PDF/A)
 
-### Change
+### Changed
 
 * Default language for XMP metadata from English to 'x-default'
+* [HexaPDF::Layout::ListBox] to use the style's font for drawing markers and to
+  fall back to Times and ZapfDingbats if necessary
 
 
 ## 0.37.2 - 2024-02-27